Some bugfixes and renamed Field -> FieldGroup & FieldGroup -> FieldWrapper
