--- src/js/components/Field/index.js
@@ -1,2 +1,0 @@
-import Field from './Field'
-export default Field
