+++ src/js/components/FieldWrapper/index.js
@@ -1,0 +1,2 @@
+import FieldWrapper from './FieldWrapper'
+export default FieldWrapper
